Matlab计算直线与广义椭圆的交点（Matlab calculating the intersection of a line and a generalized ellipse）
这是个改进的lineEllipse.m函数，用于计算直线与广义椭圆的交点。增加了异常参数处理，并且改进了交点有效性判断。
基于Paurakh的工作成果。具体参见：e A (2024). Intersection of Line and Generalized Ellipse (https://www.mathworks.com/matlabcentral/fileexchange/169741-intersection-of-line-and-generalized-ellipse), MATLAB Central File Exchange。
This is an improved lineEllipse function for calculating the intersection of a line and a generalized ellipse. The processing of exception parameters is increased, and the validity judgment of intersection points is improved. Based on Paurakh's work. See: e A (2024). Intersection of Line and Generalized Ellipse (https://www.mathworks.com/matlabcentral/fileexchange/169741-intersection-of-line-and-generalized-ellipse), MATLAB Central File Exchange.